// eic-website/src/common/enums.ts
export enum Status {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
}

export enum Gender {
  MALE = "MALE",
  FEMALE = "FEMALE",
}

export enum CurrencyTypes {
  BIRR = "BIRR",
  DOLLAR = "DOLLAR"
}

export enum Units {
  SQUARE_METERS = "SQUARE_METERS",
}

export enum PermitStatus {
  DRAFTED = "DRAFTED",
  ACCEPTED = "ACCEPTED",
  REVIEWED = "REVIEWED",
  VERIFIED = "VERIFIED",
  APPROVED = "APPROVED",
  RENEWED = "RENEWED",
  CANCELLED = "CANCELLED",
  SENT_COMPANY_NAME = "SENT_COMPANY_NAME",
  APPROVED_COMPANY_NAME = "APPROVED_COMPANY_NAME",
  SENT_NEW_COMPANY_NAME = "SENT_NEW_COMPANY_NAME",
  REGISTERED_COMPANY_NAME = "REGISTERED_COMPANY_NAME",
  REGISTERED_TIN_NUMBER  = "REGISTERED_TIN_NUMBER",
  SENT_COMPANY_REGISTRATION_BANK_SLIP = "SENT_COMPANY_REGISTRATION_BANK_SLIP",
  SENT_MEMORANDUM_BANK_SLIP = "SENT_MEMORANDUM_BANK_SLIP",
  SENT_MEMORANDUM_OF_ARTICLES = "SENT_MEMORANDUM_OF_ARTICLES",
  SENT_CREDIT_SERVICE_BANK_SLIP = "SENT_CREDIT_SERVICE_BANK_SLIP",
  SENT_SERVICE_FEE_BANK_SLIP = "SENT_SERVICE_FEE_BANK_SLIP",
  ACCEPTED_MEMORANDUM_OF_ARTICLES = "ACCEPTED_MEMORANDUM_OF_ARTICLES",
  ACCEPTED_COMPANY_REGISTRATION_BANK_SLIP = "ACCEPTED_COMPANY_REGISTRATION_BANK_SLIP",
  ACCEPTED_MEMORANDUM_BANK_SLIP = "ACCEPTED_MEMORANDUM_BANK_SLIP",
  ACCEPTED_CREDIT_SERVICE_BANK_SLIP = "ACCEPTED_CREDIT_SERVICE_BANK_SLIP",
  ACCEPTED_SERVICE_FEE_BANK_SLIP = "ACCEPTED_SERVICE_FEE_BANK_SLIP",
}

export enum TypeOfBusiness {
  'Private Limited Company' = 'Private Limited Company',
  'One Person Private Limited Company' = 'One Person Private Limited Company',
  'Share Company' = 'Share Company',
  'Public Enterprise' = 'Public Enterprise',
  'Cooperative Society' = 'Cooperative Society',
  'Sole Proprietorship' = 'Sole Proprietorship',
  'Civil Society Organization' = 'Civil Society Organization',
  'Other Engaging Business' = 'Other Engaging Business'
}